% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_catalogos.R
\name{.extraer_datos}
\alias{.extraer_datos}
\title{Extraer la tabla principal de una respuesta de SUGEVAL}
\usage{
.extraer_datos(
  parsed,
  preferidos = c("Catalogos", "Fondos", "BalanceGeneral", "CuentasOrden", "CuentaOrden",
    "EstadosResultados", "EstadoResultadoAcumulado", "EstadoResultadoMensual", "value")
)
}
\arguments{
\item{parsed}{Objeto ya interpretado desde JSON.}

\item{preferidos}{Nombres de campos que se deben priorizar.}
}
\value{
Un tibble, posiblemente vacío.
}
\description{
Extraer la tabla principal de una respuesta de SUGEVAL
}
\keyword{internal}
