% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_core_odata.R
\name{.odata_url}
\alias{.odata_url}
\title{Construir la URL de una función OData}
\usage{
.odata_url(metodo, params = list())
}
\arguments{
\item{metodo}{Nombre de la función publicada por SUGEVAL.}

\item{params}{Lista nombrada de especificaciones `valor` y `tipo`.}
}
\value{
URL completa.
}
\description{
Construir la URL de una función OData
}
\keyword{internal}
