% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_core_odata.R
\name{.odata_valor}
\alias{.odata_valor}
\title{Formatear un valor según la sintaxis OData de SUGEVAL}
\usage{
.odata_valor(x, tipo = c("string", "date", "int", "null"))
}
\arguments{
\item{x}{Valor escalar.}

\item{tipo}{Uno de `"string"`, `"date"`, `"int"` o `"null"`.}
}
\value{
Texto listo para incorporarse a la llamada OData.
}
\description{
Formatear un valor según la sintaxis OData de SUGEVAL
}
\keyword{internal}
