% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_core_odata.R
\name{.sugeval_get}
\alias{.sugeval_get}
\title{Realizar una petición GET al servicio de SUGEVAL}
\usage{
.sugeval_get(metodo, params = list(), verbose = FALSE)
}
\arguments{
\item{metodo}{Nombre de la función OData.}

\item{params}{Lista de parámetros para [`.odata_url()`].}

\item{verbose}{Si es `TRUE`, muestra la URL consultada (nunca el token).}
}
\value{
Respuesta JSON convertida a objetos de R.
}
\description{
Realizar una petición GET al servicio de SUGEVAL
}
\keyword{internal}
