% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_catalogos.R
\name{listar_puestos_bolsa}
\alias{listar_puestos_bolsa}
\title{Listar puestos de bolsa}
\usage{
listar_puestos_bolsa(estado = NULL, verbose = FALSE)
}
\arguments{
\item{estado}{`"ACTIVO"`, `"INACTIVO"` o `NULL` para no filtrar.}

\item{verbose}{Si es `TRUE`, muestra la URL consultada.}
}
\value{
Un tibble con los puestos de bolsa.
}
\description{
Listar puestos de bolsa
}
\examples{
\dontrun{
listar_puestos_bolsa("ACTIVO")
}
}
