% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_catalogos.R
\name{listar_safis}
\alias{listar_safis}
\title{Listar sociedades administradoras de fondos de inversión}
\usage{
listar_safis(estado = NULL, verbose = FALSE)
}
\arguments{
\item{estado}{`"ACTIVO"`, `"INACTIVO"` o `NULL` para no filtrar.}

\item{verbose}{Si es `TRUE`, muestra la URL consultada.}
}
\value{
Un tibble con las SAFI.
}
\description{
Listar sociedades administradoras de fondos de inversión
}
\examples{
\dontrun{
listar_safis("ACTIVO")
}
}
